backend/requirements.txt
async-timeout==4.0.3
blinker==1.6.2
caching==0.1.dev8
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
flask==2.3.3
flask-core==2.9.0
Flask-Cors==4.0.0
flask-sqlalchemy==3.1.1
future==0.18.3
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyserial==3.5
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
serial==0.0.97
SQLAlchemy==2.0.23
typing-extensions==4.8.0
urllib3==2.0.7
werkzeug==2.3.7
zipp==3.17.0